criando o metodo Save dentro da classe ComputerRepository
